+++ DiscordBot/Services/Casino/CasinoService.cs
@@ -39,7 +39,7 @@ public class CasinoService
             };
 
             var createdUser = await _databaseService.CasinoQuery.InsertCasinoUser(newUser);
-            await RecordTransaction(userId, (long)_settings.CasinoStartingTokens, TransactionType.TokenInitialisation);
+            await RecordTransaction(userId, _settings.CasinoStartingTokens, TransactionType.TokenInitialisation);
             await _loggingService.LogChannelAndFile($"{ServiceName}: Created new casino user {userId} with {_settings.CasinoStartingTokens} starting tokens");
             return createdUser;
         }
@@ -51,7 +51,7 @@ public class CasinoService
         }
     }
 
-    public async Task<bool> TransferTokens(string fromUserId, string toUserId, ulong amount)
+    public async Task<bool> TransferTokens(string fromUserId, string toUserId, long amount)
     {
         var fromUser = await GetOrCreateCasinoUser(fromUserId);
         var toUser = await GetOrCreateCasinoUser(toUserId);
@@ -64,11 +64,11 @@ public class CasinoService
         await _databaseService.CasinoQuery.UpdateTokens(toUserId, toUser.Tokens + amount, DateTime.UtcNow);
 
         // Record transactions
-        await RecordTransaction(fromUserId, -(long)amount, TransactionType.Gift, new Dictionary<string, string>
+        await RecordTransaction(fromUserId, -amount, TransactionType.Gift, new Dictionary<string, string>
         {
             ["to"] = toUserId,
         });
-        await RecordTransaction(toUserId, (long)amount, TransactionType.Gift, new Dictionary<string, string>
+        await RecordTransaction(toUserId, amount, TransactionType.Gift, new Dictionary<string, string>
         {
             ["from"] = fromUserId
         });
@@ -81,11 +81,11 @@ public class CasinoService
         try
         {
             var user = await GetOrCreateCasinoUser(userId);
-            var newBalance = (long)user.Tokens + deltaTokens;
+            var newBalance = user.Tokens + deltaTokens;
             // Prevent negative balance
             if (newBalance < 0) newBalance = 0;
 
-            await _databaseService.CasinoQuery.UpdateTokens(userId, (ulong)newBalance, DateTime.UtcNow);
+            await _databaseService.CasinoQuery.UpdateTokens(userId, newBalance, DateTime.UtcNow);
             await RecordTransaction(userId, deltaTokens, transactionType, details);
         }
         catch (Exception ex)
@@ -96,11 +96,11 @@ public class CasinoService
         }
     }
 
-    public async Task SetUserTokens(string userId, ulong amount, string adminUserId)
+    public async Task SetUserTokens(string userId, long amount, string adminUserId)
     {
         await _databaseService.CasinoQuery.UpdateTokens(userId, amount, DateTime.UtcNow);
 
-        await RecordTransaction(userId, (long)amount, TransactionType.Admin, new Dictionary<string, string>
+        await RecordTransaction(userId, amount, TransactionType.Admin, new Dictionary<string, string>
         {
             ["admin"] = adminUserId,
             ["action"] = "set"
@@ -315,7 +315,7 @@ public class CasinoService
 
     #region Daily Rewards
 
-    public async Task<(bool success, ulong tokensAwarded, ulong newBalance, DateTime nextRewardTime)> TryClaimDailyReward(string userId)
+    public async Task<(bool success, long tokensAwarded, long newBalance, DateTime nextRewardTime)> TryClaimDailyReward(string userId)
     {
         try
         {
@@ -332,7 +332,7 @@ public class CasinoService
             var tokensAwarded = _settings.CasinoDailyRewardTokens;
             var newBalance = user.Tokens + tokensAwarded;
             await _databaseService.CasinoQuery.UpdateTokensAndDailyReward(userId, newBalance, now, now);
-            await RecordTransaction(userId, (long)tokensAwarded, TransactionType.DailyReward);
+            await RecordTransaction(userId, tokensAwarded, TransactionType.DailyReward);
 
             await _loggingService.LogChannelAndFile($"{ServiceName}: User {userId} claimed daily reward of {tokensAwarded} tokens");
             return (true, tokensAwarded, newBalance, now.AddSeconds(_settings.CasinoDailyRewardIntervalSeconds));
